Split data page into a landing page and dynamic route project pages
